requirements.txt
qrcode==8.0
Pillow==10.2.0